Moved HomePage to main page
I realized that the home page should be the default page of the application so I moved it from the HomePage folder to the src directory.
